Small improvements and update.sh script

+++ update.sh
@@ -1,20 +1,4 @@
 #!/bin/bash
-# First install dotfiles 
-
-function installFiglet(){
-
-    if curl --output /dev/null --silent --head --fail  http://cygwin.mirrors.hoobly.com/x86_64/release/figlet/figlet-2.2.2-2.tar.bz2
-    then 
-        cd ../../tmp/
-        wget -O figlet.tar.bz2 http://cygwin.mirrors.hoobly.com/x86_64/release/figlet/figlet-2.2.2-2.tar.bz2
-        tar -x -C / -f figlet.tar.bz2
-        tar -tf figlet.tar.bz2 | gzip > /../etc/setup/figlet.lst.gz
-        cd ~
-    else
-        echo "Cannot install figlet! Check the url !!"
-    fi
-}
-
 
 echo "                             "
 echo "                             "
@@ -38,7 +22,7 @@ echo "|____/ \___/ \__|_| |_|_|\___||___/ "
 
 echo 
 echo 
-echo "Select what´s enviroment you are installing:"
+echo "Select whatÂ´s enviroment you are installing:"
 echo 
 echo
 echo " 1 Cygwin"
@@ -70,19 +54,22 @@ if [ $ENV_DOTFILES -eq 1 ]; then
     cp cygwin/.bash_profile ../
     cp cygwin/.bashrc ../ 
 
-    echo "Would you like to install SDKMAN(Y/n) ?"
-    echo
-    echo
-    read ENV_SDKMAN
+    if [ ! -d "${HOME}/.sdkman" ]; then
+    
+        echo "Would you like to install SDKMAN(Y/n) ?"
+        echo
+        echo
+        read ENV_SDKMAN
 
-    if [ ENV_SDKMAN = "y" ]; then
-        # Install SDKMAN
-        curl -s "https://get.sdkman.io" | bash
-        source "$HOME/.sdkman/bin/sdkman-init.sh"
-        # Install maven 
-        sdk install maven 3.6.0
-    fi
+        if [ ENV_SDKMAN = "y" ]; then
+            # Install SDKMAN
+            curl -s "https://get.sdkman.io" | bash
+            source "$HOME/.sdkman/bin/sdkman-init.sh"
+            # Install maven 
+            sdk install maven 3.6.0
+        fi
+    fi 
+fi
 
-    echo
-    echo "Sucessfuly !!!"
-fi 
+echo
+echo "Sucessfuly !!!"
